fixes navigation styling, consolidates styling into single file, adds settings page, adds logic for auto logout ever 2 minutes
