+++ KaravanProject/src/screens/Walks.js
@@ -11,6 +11,8 @@ import {
   MapView
 } from 'react-native';
 
+import styles from '../style/style.js';
+
 export default class Walks extends Component {
     render() {
     return (
@@ -19,76 +21,32 @@ export default class Walks extends Component {
           <Text style={styles.title}> Walks </Text>
         </View>
           <View style={styles.navigation}>
-            <Button 
+            <View style={styles.navigationButtonContainer}>
+              <Button 
+                color='#ffffff'
                 onPress={this.props.navigateButton.bind(this, "Home")}
                 title="Home"
-                color="#ffffff"
-                accessibilityLabel="Next Screen"
+                accessibilityLabel="Home"
               />
-            <Button 
-            onPress={this.props.navigateButton.bind(this, "Walks")}
-            title="Walks"
-            color="#ffffff"
-            accessibilityLabel="Next Screen"
-          />
-          <Button 
-            onPress={this.props.navigateButton.bind(this, "Map")}
-            title="Settings"
-            color="#ffffff"
-            accessibilityLabel="Next Screen"
-          />
+            </View>
+            <View style={styles.navigationButtonContainer}>
+              <Button
+                color='#ffffff'
+                onPress={this.props.navigateButton.bind(this, "Walks")}
+                title="Walks"
+                accessibilityLabel="Walks"
+              />
+            </View>
+            <View style={styles.navigationButtonContainer}>
+              <Button
+                color='#ffffff'
+                onPress={this.props.navigateButton.bind(this, "Settings")}
+                title="Settings"
+                accessibilityLabel="Settings"
+              />
+            </View>
           </View>
       </View>
     );
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    //justifyContent: 'center',
-    //alignItems: 'center',
-    backgroundColor: '#2980b9',
-  },
-  navigation:{
-    bottom: 0,
-    flexDirection: 'row',
-    alignItems: 'baseline',
-    justifyContent: 'space-between',
-    paddingHorizontal: 15,
-  },
-  banner: {
-    //flex: 1,
-    height: 60,
-    paddingVertical: 5,
-    flexDirection: 'row',
-    alignItems: 'flex-end',
-    justifyContent: 'center',
-    backgroundColor: '#18cd9c',
-  },
-  title: {
-    fontSize: 25,
-    fontWeight: 'bold',
-    textAlign: 'auto',
-      color: 'white'
-  },
-  map: {
-    flex: 1,
-    alignSelf: 'stretch',
-    //marginVertical: 30
-  },
-  loginButtonContainer: {
-    backgroundColor: '#1db0a2',
-    borderRadius: 5,
-    alignSelf: 'stretch',
-    paddingVertical: 10,
-    shadowColor: '#000000',
-    shadowOffset: {
-      width: 0,
-      height: 3
-    },
-    shadowRadius: 10,
-    shadowOpacity: 0.25,
-    margin: 30
-  },
-});
